Move ChangeLogPipeline and related types to the "Pipeline" namespace
