+++ src/ChangeLog.Test/Integrations/GitHub/GitHubLinkTaskTest.cs
@@ -8,7 +8,7 @@ using Grynwald.ChangeLog.Git;
 using Grynwald.ChangeLog.Integrations.GitHub;
 using Grynwald.ChangeLog.Model;
 using Grynwald.ChangeLog.Model.Text;
-using Grynwald.ChangeLog.Tasks;
+using Grynwald.ChangeLog.Pipeline;
 using Grynwald.ChangeLog.Test.Configuration;
 using Grynwald.ChangeLog.Test.Git;
 using Microsoft.Extensions.Logging;
