+++ src/ChangeLog.Test/Tasks/RenderTemplateTaskTest.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using Grynwald.ChangeLog.Configuration;
 using Grynwald.ChangeLog.Model;
+using Grynwald.ChangeLog.Pipeline;
 using Grynwald.ChangeLog.Tasks;
 using Grynwald.ChangeLog.Templates;
 using Grynwald.Utilities.IO;
