+++ src/ChangeLog.Test/Tasks/ResolveEntryReferencesTaskTest.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using Grynwald.ChangeLog.ConventionalCommits;
 using Grynwald.ChangeLog.Model;
 using Grynwald.ChangeLog.Model.Text;
+using Grynwald.ChangeLog.Pipeline;
 using Grynwald.ChangeLog.Tasks;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
