+++ src/ChangeLog/Integrations/GitLab/GitLabLinkTask.cs
@@ -9,7 +9,7 @@ using Grynwald.ChangeLog.Configuration;
 using Grynwald.ChangeLog.Git;
 using Grynwald.ChangeLog.Model;
 using Grynwald.ChangeLog.Model.Text;
-using Grynwald.ChangeLog.Tasks;
+using Grynwald.ChangeLog.Pipeline;
 using Microsoft.Extensions.Logging;
 
 namespace Grynwald.ChangeLog.Integrations.GitLab
