+++ src/ChangeLog/Pipeline/ChangeLogPipelineBuilder.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using Autofac;
 using Grynwald.ChangeLog.Tasks;
 
-namespace Grynwald.ChangeLog
+namespace Grynwald.ChangeLog.Pipeline
 {
     internal class ChangeLogPipelineBuilder : IChangeLogPipelineBuilder
     {
