+++ src/ChangeLog/Pipeline/IChangeLogTask.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Grynwald.ChangeLog.Model;
 
-namespace Grynwald.ChangeLog.Tasks
+namespace Grynwald.ChangeLog.Pipeline
 {
     public interface IChangeLogTask
     {
