+++ src/ChangeLog/Tasks/FilterVersionsTask.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using Grynwald.ChangeLog.Configuration;
 using Grynwald.ChangeLog.Model;
+using Grynwald.ChangeLog.Pipeline;
 using Microsoft.Extensions.Logging;
 using NuGet.Versioning;
