+++ src/ChangeLog/Tasks/LoadCurrentVersionTask.cs
@@ -2,6 +2,7 @@
 using Grynwald.ChangeLog.Configuration;
 using Grynwald.ChangeLog.Git;
 using Grynwald.ChangeLog.Model;
+using Grynwald.ChangeLog.Pipeline;
 using Microsoft.Extensions.Logging;
 using NuGet.Versioning;
