+++ src/ChangeLog/Tasks/ParseCommitReferencesTask.cs
@@ -4,6 +4,7 @@ using System.Text.RegularExpressions;
 using Grynwald.ChangeLog.Git;
 using Grynwald.ChangeLog.Model;
 using Grynwald.ChangeLog.Model.Text;
+using Grynwald.ChangeLog.Pipeline;
 using Microsoft.Extensions.Logging;
 
 namespace Grynwald.ChangeLog.Tasks
