+++ src/ChangeLog/Tasks/ResolveEntryReferencesTask.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Grynwald.ChangeLog.Model;
 using Grynwald.ChangeLog.Model.Text;
+using Grynwald.ChangeLog.Pipeline;
 using Microsoft.Extensions.Logging;
 
 namespace Grynwald.ChangeLog.Tasks
